+++ src/pop/integrations/langchain/__init__.py
@@ -14,16 +14,16 @@ from pop.integrations.langchain.bindings import (
 )
 
 __all__ = [
+    "create_langchain_execution_bundle",
+    "create_langchain_create_agent_kwargs",
+    "create_langchain_context_bundle",
+    "create_langchain_middleware_bundle",
+    "maybe_build_langchain_agent_spec",
     "bind_langchain_prompt",
     "bind_langchain_context",
     "bind_langchain_middleware",
     "build_langchain_agent_input",
-    "create_langchain_context_bundle",
-    "create_langchain_create_agent_kwargs",
     "create_langchain_agent_kwargs",
-    "create_langchain_execution_bundle",
     "create_langchain_execution_scaffold",
-    "create_langchain_middleware_bundle",
     "create_langchain_middleware_scaffold",
-    "maybe_build_langchain_agent_spec",
 ]